feat: remove countSize to beatBarsManager
