+++ src/frontend/managers/BeatBarsManager.js
@@ -6,6 +6,7 @@ import {
 import {makeAutoObservable} from "mobx";
 import Beat from "../models/Beat";
 import {createDefaultSoundObject} from "../vars/sounds/DEFAULT_SOUNDS";
+import {lcmArray} from "../utils/utils.js";
 
 export class BeatBarsManager {
 
@@ -13,12 +14,22 @@ export class BeatBarsManager {
         this._beats = [];
         this.metronomeManager = metronomeManager;
         this.generateBeats();
+        this._timeSignature = this.countSize();
         makeAutoObservable(this)
+    }
+
+    get timeSignature() {
+        return this._timeSignature;
     }
 
     get beats() {
         return this._beats;
     }
+
+    updateTimeSignature() {
+        this._timeSignature = this.countSize();
+    }
+
 
     addBeat(sound, note, noteAmount, beatIndex) {
         this._beats.push(new Beat(sound, note, noteAmount, beatIndex));
@@ -69,5 +80,35 @@ export class BeatBarsManager {
             beat.noteSettings = NOTES[newIndex];
         });
         this.metronomeManager.updateMetronome();
+    }
+
+    countSize() {
+        let beatAmount = this._beats.length;
+        let beatPattern = [];
+
+        for (let index = 0; index < beatAmount; index++) {
+            const { isTriplet, noteSize } = this._beats[index].noteSettings;
+            const noteAmount = this._beats[index].noteAmount;
+
+            for (let i = 0; i < (isTriplet ? 3 * noteAmount : noteAmount); i++) {
+                beatPattern.push(isTriplet ? noteSize * 3 / 2 : noteSize);
+            }
+        }
+
+        const denominator = lcmArray(beatPattern);
+        let numerator = 0;
+
+        for (let index = 0; index < beatAmount; index++) {
+            const noteAmount = this._beats[index].noteAmount;
+            const {isTriplet, noteSize} = this._beats[index].noteSettings;
+
+            if (isTriplet) {
+                numerator += noteAmount * 3 * (denominator / noteSize);
+            } else {
+                numerator += noteAmount * (denominator / noteSize);
+            }
+        }
+
+        return {numerator: numerator, tactSize: denominator};
     }
 }
